Added fixer to the leading/trailing whitespace rule
