+++ packages/ilib-lint/test/Rules.test.js
@@ -17,7 +17,6 @@
  * limitations under the License.
  */
 import { ResourceArray, ResourcePlural, ResourceString } from 'ilib-tools-common';
-import ResourceEdgeWhitespace from '../src/rules/ResourceEdgeWhitespace.js';
 import ResourceCompleteness from "../src/rules/ResourceCompleteness.js";
 import ResourceDNTTerms from '../src/rules/ResourceDNTTerms.js';
 
@@ -26,379 +25,6 @@ import { Result, IntermediateRepresentation, SourceFile } from 'ilib-lint-common
 const sourceFile = new SourceFile("a/b/c.xliff", {});
 
 describe("testRules", () => {
-    test("ResourceEdgeWhitespaceEdgesMatch", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.both-edges-match",
-            sourceLocale: "en-US",
-            source: "Some source string. ",
-            targetLocale: "de-DE",
-            target: "Some target string. ",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-
-        const result = rule.matchString(subject);
-        expect(result).toBe(undefined); // for a valid resource match result should not be produced
-    });
-
-    test("ResourceEdgeWhitespaceLeadingSpaceMissing", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.leading-space-missing",
-            sourceLocale: "en-US",
-            source: " some source string.",
-            targetLocale: "de-DE",
-            target: "some target string.",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // accidentally ommited space in front of target string
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toStrictEqual(
-            new Result({
-                rule,
-                severity: "error",
-                pathName: "a/b/c.xliff",
-                locale: "de-DE",
-                source: " some source string.",
-                id: "resource-edge-whitespace.leading-space-missing",
-                description: "Leading whitespace in target does not match leading whitespace in source",
-                highlight: `Source: <e0>⎵</e0>some… Target: <e1></e1>some…`,
-            })
-        );
-    });
-
-    test("ResourceEdgeWhitespaceLeadingSpaceExtra", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.leading-space-extra",
-            sourceLocale: "en-US",
-            source: "Some source string.",
-            targetLocale: "de-DE",
-            target: " Some target string.",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // accidentally added space in front of target string
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toStrictEqual(
-            new Result({
-                rule,
-                severity: "error",
-                pathName: "a/b/c.xliff",
-                locale: "de-DE",
-                source: "Some source string.",
-                id: "resource-edge-whitespace.leading-space-extra",
-                description: "Leading whitespace in target does not match leading whitespace in source",
-                highlight: `Source: <e0></e0>Some… Target: <e1>⎵</e1>Some…`,
-            })
-        );
-    });
-
-    test("ResourceEdgeWhitespaceTrailingSpaceMissing", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.trailing-space-missing",
-            sourceLocale: "en-US",
-            source: "Some source string ",
-            targetLocale: "de-DE",
-            target: "Some target string",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // accidentally ommited space in the end of target string
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toStrictEqual(
-            new Result({
-                rule,
-                severity: "error",
-                pathName: "a/b/c.xliff",
-                locale: "de-DE",
-                source: "Some source string ",
-                id: "resource-edge-whitespace.trailing-space-missing",
-                description: "Trailing whitespace in target does not match trailing whitespace in source",
-                highlight: `Source: …ring<e0>⎵</e0> Target: …ring<e1></e1>`,
-            })
-        );
-    });
-
-    test("ResourceEdgeWhitespaceTrailingSpaceExtra", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.trailing-space-extra",
-            sourceLocale: "en-US",
-            source: "Some source string.",
-            targetLocale: "de-DE",
-            target: "Some target string. ",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // accidentally added space in the end of target string
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toStrictEqual(
-            new Result({
-                rule,
-                severity: "error",
-                pathName: "a/b/c.xliff",
-                locale: "de-DE",
-                source: "Some source string.",
-                id: "resource-edge-whitespace.trailing-space-extra",
-                description: "Trailing whitespace in target does not match trailing whitespace in source",
-                highlight: `Source: …ing.<e0></e0> Target: …ing.<e1>⎵</e1>`,
-            })
-        );
-    });
-
-    test("ResourceEdgeWhitespaceTrailingSpaceExtraMore", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.trailing-space-extra-more",
-            sourceLocale: "en-US",
-            source: "Some source string ",
-            targetLocale: "de-DE",
-            target: "Some target string  ",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // accidentally added space in the end of target string
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toStrictEqual(
-            new Result({
-                rule,
-                severity: "error",
-                pathName: "a/b/c.xliff",
-                locale: "de-DE",
-                source: "Some source string ",
-                id: "resource-edge-whitespace.trailing-space-extra-more",
-                description: "Trailing whitespace in target does not match trailing whitespace in source",
-                highlight: `Source: …ring<e0>⎵</e0> Target: …ring<e1>⎵⎵</e1>`,
-            })
-        );
-    });
-
-    test("ResourceEdgeWhitespaceBothEdgesSpaceMissing", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.both-spaces-missing",
-            sourceLocale: "en-US",
-            source: " some source string ",
-            targetLocale: "de-DE",
-            target: "some target string",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // accidentally ommited space in front and in the end of target string
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toStrictEqual([
-            new Result({
-                rule,
-                severity: "error",
-                pathName: "a/b/c.xliff",
-                locale: "de-DE",
-                source: " some source string ",
-                id: "resource-edge-whitespace.both-spaces-missing",
-                description: "Leading whitespace in target does not match leading whitespace in source",
-                highlight: `Source: <e0>⎵</e0>some… Target: <e1></e1>some…`,
-            }),
-            new Result({
-                rule,
-                severity: "error",
-                pathName: "a/b/c.xliff",
-                locale: "de-DE",
-                source: " some source string ",
-                id: "resource-edge-whitespace.both-spaces-missing",
-                description: "Trailing whitespace in target does not match trailing whitespace in source",
-                highlight: `Source: …ring<e0>⎵</e0> Target: …ring<e1></e1>`,
-            }),
-        ]);
-    });
-
-    test("ResourceEdgeWhitespaceSpacesOnlyMatch", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.spaces-only-match",
-            sourceLocale: "en-US",
-            source: " ",
-            targetLocale: "de-DE",
-            target: " ",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // all-whitespace string
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toBe(undefined); // for a valid resource match result should not be produced
-    });
-
-    test("ResourceEdgeWhitespaceSpacesOnlyExtra", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.spaces-only-extra",
-            sourceLocale: "en-US",
-            source: " ",
-            targetLocale: "de-DE",
-            target: "  ",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // all-whitespace string
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toStrictEqual(
-            new Result({
-                rule,
-                severity: "error",
-                pathName: "a/b/c.xliff",
-                locale: "de-DE",
-                source: " ",
-                id: "resource-edge-whitespace.spaces-only-extra",
-                description: "Leading whitespace in target does not match leading whitespace in source",
-                highlight: `Source: <e0>⎵</e0> Target: <e1>⎵⎵</e1>`,
-            })
-        );
-    });
-
-    test("ResourceEdgeWhitespaceUndefinedSource", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.undefined-source",
-            sourceLocale: "en-US",
-            source: undefined,
-            targetLocale: "de-DE",
-            target: " ",
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // missing source
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toBe(undefined); // this rule should not process a resource where source is not a string
-    });
-
-    test("ResourceEdgeWhitespaceUndefinedTarget", () => {
-        expect.assertions(2);
-
-        const rule = new ResourceEdgeWhitespace();
-        expect(rule).toBeTruthy();
-
-        const resource = new ResourceString({
-            key: "resource-edge-whitespace.undefined-target",
-            sourceLocale: "en-US",
-            source: " ",
-            targetLocale: "de-DE",
-            target: undefined,
-            pathName: "resource-edge-whitespace-test.xliff",
-            state: "translated",
-        });
-        const subject = {
-            // missing target
-            source: resource.getSource(),
-            target: resource.getTarget(),
-            resource,
-            file: "a/b/c.xliff"
-        };
-        const result = rule.matchString(subject);
-        expect(result).toBe(undefined); // this rule should not process a resource where target is not a string
-    });
-
     test("ResourceCompletenessResourceComplete", () => {
         expect.assertions(2);
